% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reducers.R
\name{all_columns_reduced}
\alias{all_columns_reduced}
\title{Check if all variables reduced to a single composite}
\usage{
all_columns_reduced(.partition_step)
}
\arguments{
\item{.partition_step}{a \code{partition_step} object}
}
\value{
logical, \code{TRUE} or \code{FALSE}
}
\description{
Check if all variables reduced to a single composite
}
\keyword{internal}
