% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reducers.R
\name{all_done}
\alias{all_done}
\title{Mark the partition as complete to stop search}
\usage{
all_done(.partition_step)
}
\arguments{
\item{.partition_step}{a \code{partition_step} object}
}
\value{
a \code{partition_step} object
}
\description{
Mark the partition as complete to stop search
}
\keyword{internal}
