% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{append_mappings}
\alias{append_mappings}
\title{Append a new variable to mapping and filter out composite variables}
\usage{
append_mappings(.partition_step, new_x)
}
\arguments{
\item{.partition_step}{a \code{partition_step} object}

\item{new_x}{the name of the reduced variable}
}
\value{
a \code{tibble}, the mapping key
}
\description{
Append a new variable to mapping and filter out composite variables
}
\keyword{internal}
