% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{as_partition}
\alias{as_partition}
\title{Return a partition object}
\usage{
as_partition(partition_step)
}
\arguments{
\item{.partition_step}{a \code{partition_step} object}
}
\value{
a \code{partition} object
}
\description{
\code{as_partition()} is called when partitioning is complete. It scrubs a
\code{partition_step} object, cleans the reduced variable names, adds mapping
indices, and sorts the composite variables.
}
\keyword{internal}
