% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{get_indices}
\alias{get_indices}
\alias{add_indices}
\alias{sort_mapping}
\title{Process mapping key to return from \code{partition()}}
\usage{
get_indices(.partition_step)

add_indices(.partition_step)

sort_mapping(.partition_step)
}
\arguments{
\item{.partition_step}{a \code{partition_step} object}
}
\value{
a \code{partition_step} object
}
\description{
\code{add_indices()} uses \code{get_indices()} to add the variable positions to the
mapping key. \code{sort_mapping()} sorts the composite variables of each reduced
variable by their position in the original data.
}
\keyword{internal}
