% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitioners.R
\name{as_partitioner}
\alias{as_partitioner}
\title{Create a partitioner}
\usage{
as_partitioner(director, metric, reducer)
}
\arguments{
\item{director}{a function that directs, possibly created by \code{\link[=as_director]{as_director()}}}

\item{metric}{a function that measures, possibly created by \code{\link[=as_metric]{as_metric()}}}

\item{reducer}{a function that reduces, possibly created by \code{\link[=as_reducer]{as_reducer()}}}
}
\value{
a \code{partitioner}
}
\description{
Partitioners are functions that tell the partition algorithm 1)
what to try to reduce 2) how to measure how much information is lost from
the reduction and 3) how to reduce the data. In partition, functions that
handle 1) are called directors, functions that handle 2) are called
metrics,  and functions that handle 3) are called reducers. partition has a
number of pre-specificed partitioners for agglomerative data reduction.
Custom partitioners can be created with \code{\link[=as_partitioner]{as_partitioner()}}.
}
\examples{

as_partitioner(
  director = direct_distance_pearson(),
  metric = metric_icc(),
  reducer = reduce_scaled_means()
)

}
