% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reducers.R
\name{binary_k_search}
\alias{binary_k_search}
\title{Search for best \code{k} using the binary search method}
\usage{
binary_k_search(.partition_step)
}
\arguments{
\item{.partition_step}{a \code{partition_step} object}
}
\value{
a \code{partition_step} object
}
\description{
Search for best \code{k} using the binary search method
}
\keyword{internal}
