% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reducers.R
\name{build_next_name}
\alias{build_next_name}
\title{Create new variable name based on prefix and previous reductions}
\usage{
build_next_name(.partition_step)
}
\arguments{
\item{.partition_step}{a \code{partition_step} object}
}
\value{
a character vector
}
\description{
Create new variable name based on prefix and previous reductions
}
\keyword{internal}
