% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reducers.R
\name{calculate_new_variable}
\alias{calculate_new_variable}
\title{Calculate or retrieve stored reduced variable}
\usage{
calculate_new_variable(.partition_step, .f)
}
\arguments{
\item{.partition_step}{a \code{partition_step} object}
}
\value{
a numeric vector, the reduced variable
}
\description{
Calculate or retrieve stored reduced variable
}
\keyword{internal}
