% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reducers.R
\name{k_searching_forward}
\alias{k_searching_forward}
\alias{k_searching_backward}
\alias{boundary_found}
\title{Assess \code{k} search}
\usage{
k_searching_forward(.partition_step)

k_searching_backward(.partition_step)

boundary_found(.partition_step)
}
\arguments{
\item{.partition_step}{a \code{partition_step} object}
}
\value{
logical, \code{TRUE} or \code{FALSE}
}
\description{
\code{k_searching_forward()} and \code{k_searching_backward()} check the direction of
the \code{k} search metric. \code{boundary_found()} checks if the last value of \code{k} was
under the threshold while the current value is over
}
\keyword{internal}
