% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reducers.R
\name{under_threshold}
\alias{under_threshold}
\alias{above_threshold}
\alias{is_within}
\alias{metric_within_tolerance}
\title{Compare metric to threshold}
\usage{
under_threshold(.partition_step)

above_threshold(.partition_step)

is_within(.x, .y, .e)

metric_within_tolerance(.partition_step)
}
\arguments{
\item{.partition_step}{a \code{partition_step} object}
}
\value{
logical, \code{TRUE} or \code{FALSE}
}
\description{
\code{under_threshold()} and \code{above_threshold()} check relative location of the
metric. \code{metric_within_tolerance()} uses \code{is_within()} to check if the metric
is within in the range of the threshold plus/minus the tolerance.
}
\keyword{internal}
