% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directors.R
\name{direct_k_cluster}
\alias{direct_k_cluster}
\title{Target based on K-means clustering}
\usage{
direct_k_cluster(.partition_step, search = c("binary", "linear"))
}
\arguments{
\item{.partition_step}{a \code{partition_step} object}

\item{search}{The search method. Binary search is generally more efficient
but linear search can be faster in very low dimensions.}
}
\value{
a \code{partition_step} object
}
\description{
\code{direct_k_cluster()} assigns each variable to a cluster using
K-means. As the partition looks for the best reduction,
\code{direct_k_cluster()} iterates through values of \code{k} to assign clusters.
This search is handled by the binary search method by default and thus
does not necessarily need to fit every value of k.
}
