% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{direct_measure_reduce}
\alias{direct_measure_reduce}
\title{Apply a partitioner}
\usage{
direct_measure_reduce(.x, partitioner)
}
\arguments{
\item{partitioner}{a partitioner, as created from \code{\link[=as_partitioner]{as_partitioner()}}.}

\item{.partition_step}{a \code{partition_step} object}
}
\value{
a \code{partition_step} object
}
\description{
Apply a partitioner
}
\seealso{
\code{\link[=as_partitioner]{as_partitioner()}}
}
\keyword{internal}
