% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{filter_reduced}
\alias{filter_reduced}
\alias{unnest_reduced}
\title{Filter the reduced mappings}
\usage{
filter_reduced(.partition)

unnest_reduced(.partition)
}
\arguments{
\item{.partition}{a \code{partition} object}
}
\value{
a \code{tibble} with mapping key
}
\description{
\code{filter_reduced()} and \code{unnest_reduced()} are convenience functions to
quickly retrieve the mappings for only the reduced variables.
\code{filter_reduced()} returns a nested \code{tibble} while \code{unnest_reduced()} unnests
it.
}
