% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directors.R
\name{find_algorithm}
\alias{find_algorithm}
\title{Which kmeans algorithm to use?}
\usage{
find_algorithm(algorithm)
}
\arguments{
\item{algorithm}{the kmeans algorithm to use}
}
\value{
a kmeans function
}
\description{
Which kmeans algorithm to use?
}
\keyword{internal}
