% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directors.R
\name{find_min_distance_variables}
\alias{find_min_distance_variables}
\title{Find the index of the pair with the smallest distance}
\usage{
find_min_distance_variables(.x)
}
\arguments{
\item{.x}{a distance matrix}
}
\value{
a character vector with the names of the minimum pair
}
\description{
Find the index of the pair with the smallest distance
}
\keyword{internal}
