% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directors.R
\name{fit_distance_matrix}
\alias{fit_distance_matrix}
\title{Fit a distance matrix using correlation coefficients}
\usage{
fit_distance_matrix(.partition_step, spearman = FALSE)
}
\arguments{
\item{.partition_step}{a \code{partition_step} object}

\item{spearman}{Logical. Use Spearman's correlation?}
}
\value{
a \code{matrix} of size \code{p} by \code{p}
}
\description{
Fit a distance matrix using correlation coefficients
}
\keyword{internal}
