% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directors.R
\name{guess_init_k}
\alias{guess_init_k}
\title{Guess initial \code{k} based on threshold and \code{p}}
\usage{
guess_init_k(.partition_step)
}
\arguments{
\item{.partition_step}{a \code{partition_step} object}
}
\value{
an integer
}
\description{
Guess initial \code{k} based on threshold and \code{p}
}
\keyword{internal}
