% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{fill_in_missing}
\alias{fill_in_missing}
\alias{swap_nans}
\title{Process reduced variables when missing data}
\usage{
fill_in_missing(x, .na, .fill = NA)

swap_nans(.x)
}
\arguments{
\item{x}{a vector, the reduced variable}

\item{.na}{a logical vector marking which are missing}

\item{.fill}{what to fill the missing locations with}
}
\value{
a vector of length nrow(original data)

a character vector
}
\description{
Process reduced variables when missing data
}
\keyword{internal}
