% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{increase_hits}
\alias{increase_hits}
\alias{get_hits}
\title{Count and retrieve the number of metrics below threshold}
\usage{
increase_hits(.partition_step)

get_hits(.partition_step)
}
\arguments{
\item{.partition_step}{a \code{partition_step} object}
}
\description{
Count and retrieve the number of metrics below threshold
}
\keyword{internal}
