% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{icc}
\alias{icc}
\title{Calculate the intraclass correlation coefficient}
\usage{
icc(.x, method = c("r", "c"))
}
\arguments{
\item{.x}{a data set}

\item{method}{The method source: both the pure R and C++ versions are efficient}
}
\value{
a numeric vector of length 1
}
\description{
\code{icc()} efficiently calculates the ICC for a numeric data set.
}
\examples{
library(dplyr)
iris \%>\%
  select_if(is.numeric) \%>\%
  icc()
}
