% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{icc_r}
\alias{icc_r}
\title{Calculate the intraclass correlation coefficient}
\usage{
icc_r(.x)
}
\arguments{
\item{.x}{a data set}
}
\value{
a numeric vector of length 1
}
\description{
\code{icc_r()} efficiently calculates the ICC for a numeric data set in pure R.
}
\keyword{internal}
