% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{is_partition}
\alias{is_partition}
\title{Is this object a partition?}
\usage{
is_partition(x)
}
\arguments{
\item{x}{an object to be tested}
}
\value{
logical: \code{TRUE} or \code{FALSE}
}
\description{
Is this object a partition?
}
