% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{is_partition_step}
\alias{is_partition_step}
\title{Is this object a \code{partition_step}?}
\usage{
is_partition_step(x)
}
\arguments{
\item{x}{an object to be tested}
}
\value{
logical: \code{TRUE} or \code{FALSE}
}
\description{
Is this object a \code{partition_step}?
}
