% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitioners.R
\name{is_partitioner}
\alias{is_partitioner}
\title{Is this object a partitioner?}
\usage{
is_partitioner(x)
}
\arguments{
\item{x}{an object to be tested}
}
\value{
logical: \code{TRUE} or \code{FALSE}
}
\description{
Is this object a partitioner?
}
