% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{is_same_function}
\alias{is_same_function}
\title{Are two functions the same?}
\usage{
is_same_function(x, y)
}
\arguments{
\item{x, y}{functions to compare}
}
\value{
logical: \code{TRUE} or \code{FALSE}
}
\description{
\code{is_same_function()} compares functions correctly even if they are partialized.
}
\keyword{internal}
