% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directors.R
\name{k_exhausted}
\alias{k_exhausted}
\title{Have all values of \code{k} been checked for metric?}
\usage{
k_exhausted(.partition_step)
}
\arguments{
\item{.partition_step}{a \code{partition_step} object}
}
\value{
logical: \code{TRUE} or \code{FALSE}
}
\description{
Have all values of \code{k} been checked for metric?
}
\keyword{internal}
