% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directors.R
\name{find_algorithm}
\alias{find_algorithm}
\alias{kmean_assignment_c}
\alias{kmean_assignment_r}
\title{Which kmeans algorithm to use?}
\usage{
find_algorithm(algorithm, seed)

kmean_assignment_c(.data, k, n_iter = 10L, verbose = FALSE, seed = 1L)

kmean_assignment_r(.data, k, algorithm = "Hartigan-Wong", seed = 1L)
}
\arguments{
\item{algorithm}{the kmeans algorithm to use}
}
\value{
a kmeans function
}
\description{
\code{find_algorithm()} returns a function to assign k-means cluster.
\code{kmean_assignment_r()} wraps around \code{\link[=kmeans]{kmeans()}} to pull the correct
assignments.
}
\keyword{internal}
