% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reducers.R
\name{linear_k_search}
\alias{linear_k_search}
\title{Search for best \code{k} using the linear search method}
\usage{
linear_k_search(.partition_step, n_hits = 4)
}
\arguments{
\item{.partition_step}{a \code{partition_step} object}
}
\value{
a \code{partition_step} object
}
\description{
Search for best \code{k} using the linear search method
}
\keyword{internal}
