% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{mappings}
\alias{mappings}
\alias{mapping_groups}
\title{Return partition mapping key}
\usage{
mappings(.partition)

mapping_groups(.partition, indices = FALSE)
}
\arguments{
\item{.partition}{a \code{partition} object}

\item{indices}{logical. Return just the indices instead of the names? Default is \code{FALSE}.}
}
\value{
a \code{tibble}
}
\description{
\code{mappings()} returns a tidy data frame with each reduced variable and its
mapping and information loss. \code{mapping_groups()} returns a list of mappings
(either the variable names or their column position).
}
\examples{


}
