% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directors.R
\name{matrix_is_exhausted}
\alias{matrix_is_exhausted}
\title{Have all pairs of variables been checked for metric?}
\usage{
matrix_is_exhausted(.partition_step)
}
\arguments{
\item{.partition_step}{a \code{partition_step} object}
}
\value{
logical: \code{TRUE} or \code{FALSE}
}
\description{
Have all pairs of variables been checked for metric?
}
\keyword{internal}
