% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{measure_std_mutualinfo}
\alias{measure_std_mutualinfo}
\title{Measure the information loss of reduction using standardized mutual
information}
\usage{
measure_std_mutualinfo(.partition_step)
}
\arguments{
\item{.partition_step}{a \code{partition_step} object}
}
\value{
a \code{partition_step} object
}
\description{
Metrics are functions that tell how much information would be
lost for a given reduction in the data. reduce.  \code{\link[=as_measure]{as_measure()}} is a
helper function to create new metrics to be used in \code{partitioner}s.
\code{partitioner}s can be created with \code{\link[=as_partitioner]{as_partitioner()}}.

\code{measure_std_mutualinfo()} assesses information loss by
calculating the standardized mutual information for the target variables.
See \code{\link[=mutual_information]{mutual_information()}}.
}
\seealso{
Other metrics: 
\code{\link{as_measure}()},
\code{\link{measure_icc}()},
\code{\link{measure_min_icc}()},
\code{\link{measure_min_r2}()},
\code{\link{measure_variance_explained}()}
}
\concept{metrics}
