% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{measure_variance_explained}
\alias{measure_variance_explained}
\title{Measure the information loss of reduction using the variance explained}
\usage{
measure_variance_explained(.partition_step)
}
\arguments{
\item{.partition_step}{a \code{partition_step} object}
}
\value{
a \code{partition_step} object
}
\description{
Metrics are functions that tell how much information would be
lost for a given reduction in the data. reduce.  \code{\link[=as_measure]{as_measure()}} is a
helper function to create new metrics to be used in \code{partitioner}s.
\code{partitioner}s can be created with \code{\link[=as_partitioner]{as_partitioner()}}.

\code{measure_variance_explained()} assesses information loss by calculating the
variance explained by the first component of a principal components analysis.
}
\seealso{
Other metrics: 
\code{\link{as_measure}()},
\code{\link{measure_icc}()},
\code{\link{measure_min_icc}()},
\code{\link{measure_min_r2}()},
\code{\link{measure_std_mutualinfo}()}
}
\concept{metrics}
