% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metric_icc}
\alias{metric_icc}
\title{Measure the information loss of reduction using intraclass correlation
coefficient}
\usage{
metric_icc(.partition_step)
}
\arguments{
\item{.partition_step}{a \code{partition_step} object}
}
\value{
a \code{partition_step} object
}
\description{
\code{metric_icc()} assesses information loss by calculating the
intraclass correlation coefficient for the target variables.
}
