% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metric_min_icc}
\alias{metric_min_icc}
\title{Measure the information loss of reduction using the minimum intraclass
correlation coefficient}
\usage{
metric_min_icc(.partition_step, search_method = c("binary", "linear"))
}
\arguments{
\item{.partition_step}{a \code{partition_step} object}

\item{search_method}{The search method. Binary search is generally more efficient
but linear search can be faster in very low dimensions.}
}
\value{
a \code{partition_step} object
}
\description{
\code{metric_min_icc()} assesses information loss by calculating the
intraclass correlation coefficient for each set of the target variables and
finding their minimum.
}
