% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metric_min_r2}
\alias{metric_min_r2}
\title{Measure the information loss of reduction using minimum R-squared}
\usage{
metric_min_r2(.partition_step)
}
\arguments{
\item{.partition_step}{a \code{partition_step} object}
}
\value{
a \code{partition_step} object
}
\description{
\code{metric_variance_explained()} assesses information loss by
calculating the minimum R-squared for the target variables.
}
