% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metric_variance_explained}
\alias{metric_variance_explained}
\title{Measure the information loss of reduction using the variance explained}
\usage{
metric_variance_explained(.partition_step)
}
\arguments{
\item{.partition_step}{a \code{partition_step} object}
}
\value{
a \code{partition_step} object
}
\description{
\code{metric_variance_explained()} assesses information loss by calculating the
variance explained by the first component of a principal components analysis.
}
