% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutual_information.R
\name{mutual_information}
\alias{mutual_information}
\title{Calculate the standardized mutual information of a data set}
\usage{
mutual_information(.data)
}
\arguments{
\item{.data}{a dataframe of numeric values}
}
\value{
a list containing the standardized MI and the scaled row means
}
\description{
\code{mutual_information} calculate the standardized mutual information of a data
set using the \code{infotheo} package.
}
\examples{
library(dplyr)
iris \%>\%
  select_if(is.numeric) \%>\%
  mutual_information()

}
