% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params-data.R
\docType{data}
\name{params}
\alias{params}
\title{Dataset of estimated negative binomial zero inflation parameters from count data
from Dr. Wendy Cozen led study on microbiome and colon polyps}
\format{A list object, with each element a vector of parameters including,
NB intercept, NB log(theta), ZIN intercept, ZIN intercept p-value}
\usage{
data(params)
}
\description{
Data parameters were generated from genus summarized 16s rRNA count data
adjusted for read depth, excluding taxa with > 80 percent zeros.
Samples included > 20,000 reads
NBZIN parameters estimated using the zeroinfl() function of the pscl R package.
}
\examples{
data(params)
}
\keyword{datasets}
