% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitioners.R
\name{part_icc}
\alias{part_icc}
\title{Partitioner: distance, ICC, scaled means}
\usage{
part_icc(spearman = FALSE)
}
\arguments{
\item{spearman}{logical. Use Spearman's correlation for distance matrix?}
}
\value{
a \code{partitioner}
}
\description{
\code{part_icc()} uses the following functions:
\itemize{
\item \strong{director}: \code{direct_distance()}, Miniumum Distance
\item \strong{metric}: \code{metric_icc()}, Intraclass Correlation
\item \strong{reducer}: \code{reduce_scaled_mean()}, Scaled Row Means

@md
}
}
