% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitioners.R
\name{part_kmeans}
\alias{part_kmeans}
\title{Partitioner: K-means, ICC, scaled means}
\usage{
part_kmeans(algorithm = c("armadillo", "Hartigan-Wong", "Lloyd", "Forgy",
  "MacQueen"), search = c("binary", "linear"), init_k = NULL)
}
\arguments{
\item{algorithm}{The K-Means algorithm to use. The default is a fast version
of the LLoyd algorithm written in armadillo. The rest are options in
\code{\link[=kmeans]{kmeans()}}. In general, armadillo is fastest, but the other algorithms can
be faster in high dimensions.}

\item{search}{The search method. Binary search is generally more efficient
but linear search can be faster in very low dimensions.}

\item{init_k}{The initial k to test. If \code{NULL}, then the initial k is the
threshold times the number of variables.}
}
\value{
a \code{partitioner}
}
\description{
\code{part_minr2()} uses the following functions:
\itemize{
\item \strong{director}: \code{direct_k_cluster()}, K-Means Clusters
\item \strong{metric}: \code{metric_min_icc()}, Minimum Intraclass Correlation
\item \strong{reducer}: \code{reduce_kmeans()}, Scaled Row Means

@md
}
}
