% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitioners.R
\name{part_pc1}
\alias{part_pc1}
\title{Partitioner: distance, first principal component, scaled means}
\usage{
part_pc1(spearman = FALSE)
}
\arguments{
\item{spearman}{logical. Use Spearman's correlation for distance matrix?}
}
\value{
a \code{partitioner}
}
\description{
\code{part_minr2()} uses the following functions:
\itemize{
\item \strong{director}: \code{direct_distance()}, Miniumum Distance
\item \strong{metric}: \code{metric_variance_explained()}, Variance Explained (PCA)
\item \strong{reducer}: \code{reduce_scaled_mean()}, Scaled Row Means

@md
}
}
