% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partdata.R
\docType{data}
\name{partdata}
\alias{partdata}
\title{320 instances of the partition problem.}
\format{A list with 2 elements:
\describe{
  \item{instances}{a list with 320 integer sets (numeric vector)}
  \item{optimums}{a numeric vector of the 320 optimums}
}}
\usage{
partdata
}
\description{
The instances contains 10, 20, 25, or 30 integers.
They have been optimally solved using lpSolve.
}
\keyword{datasets}

