% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartEval.R
\name{PartEval}
\alias{PartEval}
\title{Evaluate the objective value of the partition}
\usage{
PartEval(sizes, part)
}
\arguments{
\item{sizes}{a numeric vector which represents the integers (>0) in the set S.}

\item{part}{a logical vector which represents the indicator function of S1 in the partition S=S1+S2.}
}
\value{
the absolute difference between the sum of integers in S1 and S2.
}
\description{
Evaluate the objective value of the partition
}
\examples{
sizes <- c(13, 11, 8, 20, 20, 11, 16, 17, 4, 13)
part <-  c(TRUE, TRUE, FALSE, FALSE, TRUE, FALSE, FALSE, FALSE, TRUE, TRUE)
PartEval(sizes, part)
}

