% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartGenInt.R
\name{PartGenInt}
\alias{PartGenInt}
\title{Generate a list of partition problem's instances}
\usage{
PartGenInt(m = 1, cardinalities = 10, precisions = 100, modes = 0)
}
\arguments{
\item{m}{number of integer sets for each parameter}

\item{cardinalities}{cardinalities of the integer sets.}

\item{precisions}{maximum values of an integer. The minimal value is 1.}

\item{modes}{a distribution mode (numeric vector) or a list of distribution modes (multi-modal distribution).}
}
\value{
a list of partition problem's instances (numeric vector of strictly positive integers). The list is changed into a vector if its length is 1.
}
\description{
For each parameter, generate m sets of integer.
The parameters are the cartesian product of the \code{cardinalities}, the \code{precisions}, and the multi-modal distributions \code{modes}.
}
\details{
The mode of an integer is drawn with a a uniform distribution.
A mode is an integer such that :
 0 <= mode  : uniform distribution
 0< mode <1 : binomial with probability of success 'mode'
 mode >= 1  : binomial with random probability of success
}
\examples{
PartGenInt()
PartGenInt( m = 2, cardinalities = c(5,10), precisions = c(10, 100))
PartGenInt( modes = list(0, 1, c(0.2, 0.8)))
}

