% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartGenTodd.R
\name{PartGenTodd}
\alias{PartGenTodd}
\title{Generate Todd Subset Sum problem's instances.}
\usage{
PartGenTodd(n)
}
\arguments{
\item{n}{numeric vector of integer set cardinalities}
}
\value{
a list of integer sets (numeric vector).
}
\description{
Generate Todd Subset Sum problem's instances.
}
\section{References}{

See Knapsack Problems, Sylvano Martello and Paolo Toth, chapter 4, p. 128.
}
\examples{
PartGenTodd(5)
PartGenTodd(2:4)
}

