% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartIsPerfect.R
\name{PartIsPerfect}
\alias{PartIsPerfect}
\title{Check if the partition is perfect.}
\usage{
PartIsPerfect(sizes, part)
}
\arguments{
\item{sizes}{a numeric vector which represents the integers (>0) in the set S.}

\item{part}{a logical vector which represents the indicator function of S1 in the partition S=S1+S2.}
}
\value{
\code{true} if the partition is perfect.
}
\description{
A partition is perfect if the difference between the sum of integers of S1 and S2 is equal to 0 (resp. 1) for a even (resp. odd) set.
}
\examples{
PartIsPerfect(seq(4), c(FALSE,TRUE,TRUE,FALSE))
PartIsPerfect(seq(4), c(TRUE,TRUE,TRUE,FALSE))
}

