% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition-package.R
\docType{package}
\name{partition-package}
\alias{partition-package}
\title{Agglomerative partitioning for dimension reduction}
\description{
This package implements a formal statistic
}
\details{
\tabular{ll}{ Package: \tab cit\cr Type: \tab Package\cr Version: \tab
1.0\cr Date: \tab 2017-12-1\cr License: \tab Artistic-2.0\cr LazyLoad: \tab
yes\cr } This package implements
}
\references{
Millstein J,
}
\author{
Joshua Millstein

Maintainer: Joshua Millstein \href{mailto:joshua.millstein@usc.edu}{joshua.millstein@usc.edu} Joshua Millstein
}
\keyword{clustering}
\keyword{dimension}
\keyword{reduction}
