% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{partition}
\alias{partition}
\title{Agglomerative partitioning dimension reduction}
\usage{
partition(.data, threshold, partitioner = part_icc(), tolerance = 0.01,
  niter = NULL, x = "reduced_var", .sep = "_")
}
\arguments{
\item{.data}{a data.frame to partition}

\item{threshold}{the minimum information loss acceptable}

\item{partitioner}{a \code{partitioner}. See the \code{part_*()} functions and
\code{\link[=as_partitioner]{as_partitioner()}}.}

\item{tolerance}{a small tolerance within the threshold; if a reduction is
within the threshold plus/minus the tolerance, it will reduce.}

\item{niter}{the number of iterations. By default, it is calculated as 20% of
the number of variables or 10, whichever is larger.}

\item{x}{the prefix of the new variable names}

\item{.sep}{a character vector that separates \code{x} from the number (e.g. "reduced_var_1").}
}
\value{
a \code{partition} object
}
\description{
Agglomerative partitioning dimension reduction
}
