% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{partition_scores}
\alias{partition_scores}
\alias{fitted.partition}
\title{Return the reduced data from a partition}
\usage{
partition_scores(object, ...)

\method{fitted}{partition}(object, ...)
}
\arguments{
\item{object}{a \code{partition} object}

\item{...}{not currently used (for S3 consistency with \code{fitted()})}
}
\value{
a tibble containing the reduced data for the partition
}
\description{
Return the reduced data from a partition
}
