% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartLpSolve.R
\name{PartLpSolve}
\alias{PartLpSolve}
\title{Solve a subset sum integer program that models the partition problem's instance using lpsolve.}
\usage{
PartLpSolve(sizes)
}
\arguments{
\item{sizes}{a numeric vector which represents the integers (>0) in the set S.}
}
\description{
Symmetry breaking : the sum of integers in S1 is lower than the one in S_2.
WARNING: the timeout does not work. It does not interrupt the lpsolve subprocess. 
WARNING: the results are wrong for high precision integers.
}
\examples{
sizes <- c(13, 11, 8, 20, 20, 11, 16, 17, 4, 13)
part <-  PartLpSolve(sizes)
PartPrint(sizes, part)
}

