% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartPrint.R
\name{PartPrint}
\alias{PartPrint}
\title{Pretty print and plot (optional) the partition}
\usage{
PartPrint(sizes, part, plot = FALSE)
}
\arguments{
\item{sizes}{a numeric vector which represents the integers (>0) in the set S.}

\item{part}{a logical vector which represents the indicator function of S1 in the partition S=S1+S2.}

\item{plot}{barplot of the partition}
}
\description{
Pretty print and plot (optional) the partition
}
\examples{
sizes <- c(13, 11, 8, 20, 20, 11, 16, 17, 4, 13)
part <-  c(TRUE, TRUE, FALSE, FALSE, TRUE, FALSE, FALSE, FALSE, TRUE, TRUE)
PartPrint(sizes, part)
}

