% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartRace.R
\name{PartRace}
\alias{PartRace}
\title{Race to solve the partition problem}
\usage{
PartRace(candidates, tasks, maxExp = 0, stat.test = c("friedman",
  "t.bonferroni", "t.holm", "t.none"), conf.level = 0.95, first.test = 10,
  interactive = TRUE, log.file = "", timeout = NA, plot.file = "")
}
\arguments{
\item{candidates}{a character vector which contains the names of the solver's functions}

\item{tasks}{a list of integer sets (numeric vector)}

\item{maxExp}{Maximum number of experiments (i.e. evaluations of the
    function \code{race.wrapper}) that are allowed before selecting the
    best candidate. If \code{maxExp=0}, no limit is imposed\ldots very
    unrealistic in practice.}

\item{stat.test}{Statistical test to be used for discarding inferior
    candidates.}

\item{conf.level}{The confidence level to be used for the statistical
    test.}

\item{first.test}{The first test for discarding inferior candidates is
    performed only when all candidates have been evaluated on a minimum
    number of tasks equal to \code{first.test}.}

\item{interactive}{If \code{TRUE}, print a progress report on the
    standard output.}

\item{log.file}{File for saving periodically the state of the race.}

\item{timeout}{per run in seconds (one candidate executes one task) that requires the package \pkg{R.utils}}

\item{plot.file}{filename to plot the progress of the race.}
}
\description{
Set up, execute, and log a race between several solvers (\code{candidates}) to solve partition problem's instances (\code{task}).
It requires the package \pkg{race}.
}
\examples{
## Generate five random instances of the partition problem
tasks <- PartGenInt( m = 5)
## Functions that return a random partition 
BeatMe <- function(sizes) sample(c(TRUE,FALSE),length(sizes), replace=TRUE)
BeatMeAgain <- BeatMe
## vector of candidate names
candidates <- c("BeatMe", "BeatMeAgain")
## Race between the two functions/algorithms
PartRace(candidates, tasks, plot.file = NA)
}
\seealso{
\link[race]{race}, \link[race]{race.wrapper},
}

