% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{paste_director}
\alias{paste_director}
\alias{paste_metric}
\alias{paste_reducer}
\title{Lookup partitioner types to print in English}
\usage{
paste_director(x)

paste_metric(x)

paste_reducer(x)
}
\arguments{
\item{x}{the function for which to find a description}
}
\value{
a description of the parts of the partitioner
}
\description{
Lookup partitioner types to print in English
}
\keyword{internal}
