% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{cat_bold}
\alias{cat_bold}
\alias{cat_white}
\alias{cat_subtle}
\alias{paste_subtle}
\title{Print to the console in color}
\usage{
cat_bold(...)

cat_white(...)

cat_subtle(...)

paste_subtle(...)
}
\arguments{
\item{...}{text to print. Passed to \code{cat()} or \code{paste()}.}
}
\description{
Print to the console in color
}
\keyword{internal}
