% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{pull_composite_variables}
\alias{pull_composite_variables}
\alias{expand_mappings}
\alias{get_names}
\title{Access mapping variables}
\usage{
pull_composite_variables(.partition_step)

expand_mappings(x, .mapping_key)

get_names(.partition_step, target_list)
}
\arguments{
\item{.partition_step}{a \code{partition_step} object}

\item{.mapping_key}{a mapping key}

\item{target_list}{a list of composite variables}
}
\value{
a vector containing mappings
}
\description{
\code{pull_composite_variables()} takes a target and finds all the composite
variables (e.g. if a reduced variable is a target, it finds all the variables
the reduced variable is created from). \code{expand_mappings()} extracts the
composite variables of a given variable. \code{get_names()} finds the variable
names for a list of column positions.
}
\keyword{internal}
