% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reducers.R
\name{reduce_first_component}
\alias{reduce_first_component}
\title{Reduce selected variables to first principal component}
\usage{
reduce_first_component(.partition_step)
}
\arguments{
\item{.partition_step}{a \code{partition_step} object}
}
\value{
a \code{partition_step} object
}
\description{
\code{reduce_first_component()} returns the first component from the
principal components analysis of the target variables.
}
