% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reducers.R
\name{reduce_kmeans}
\alias{reduce_kmeans}
\title{Reduce selected variables to scaled means}
\usage{
reduce_kmeans(.partition_step, search = c("binary", "linear"))
}
\arguments{
\item{.partition_step}{a \code{partition_step} object}

\item{search}{The search method. Binary search is generally more efficient
but linear search can be faster in very low dimensions.}
}
\value{
a \code{partition_step} object
}
\description{
\code{reduce_kmeans()} is efficient in that it doesn't reduce until
the closest \code{k} to the information threshold is found.
}
