% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{reduce_mappings}
\alias{reduce_mappings}
\title{Create a mapping key out of a list of targets}
\usage{
reduce_mappings(.partition_step, target_list)
}
\arguments{
\item{.partition_step}{a \code{partition_step} object}

\item{target_list}{a list of composite variables}
}
\value{
a \code{tibble}, the mapping key
}
\description{
Create a mapping key out of a list of targets
}
\keyword{internal}
