% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reducers.R
\name{reduce_scaled_mean}
\alias{reduce_scaled_mean}
\title{Reduce selected variables to scaled means}
\usage{
reduce_scaled_mean(.partition_step)
}
\arguments{
\item{.partition_step}{a \code{partition_step} object}
}
\value{
a \code{partition_step} object
}
\description{
Reducers are functions that tell the partition algorithm how
to reduce the data. \code{\link[=as_reducer]{as_reducer()}} is a helper function to create new
reducers to be used in \code{partitioner}s. \code{partitioner}s can be created with
\code{\link[=as_partitioner]{as_partitioner()}}.

\code{reduce_scaled_mean()} returns the scaled row means of the
target variables to reduce.
}
\seealso{
Other reducers: 
\code{\link{as_reducer}()},
\code{\link{reduce_first_component}()},
\code{\link{reduce_kmeans}()}
}
\concept{reducers}
