% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reducers.R
\name{reduce_data}
\alias{reduce_data}
\alias{map_data}
\title{Reduce a target}
\usage{
reduce_data(.partition_step, .f, first_match = TRUE)

map_data(.partition_step, .f, first_match = FALSE)
}
\arguments{
\item{.partition_step}{a \code{partition_step} object}

\item{.f}{a function to reduce the data to either a vector or a data.frame}

\item{first_match}{logical. Should the partition algorithm stop when it finds
a reduction that is equal to the threshold? Default is \code{TRUE} for reducers
that return a \code{data.frame} and \code{FALSE} for reducers that return a vector}
}
\value{
a \code{partition_step} object
}
\description{
\code{reduce_data()} and \code{map_data()} apply the data reduction to the targets
found in the director step. They only do so if the metric is above the
threshold, however. \code{reduce_data()} is for functions that return vectors
while \code{map_data()} is for functions that return \code{data.frames}. If you're
using \code{\link[=as_reducer]{as_reducer()}}, there's no need to call these functions directly.
}
\examples{

reduce_row_means <- function(.partition_step, .data) {
  reduce_data(.partition_step, rowMeans)
}

replace_partitioner(
  part_icc,
  reducer = reduce_row_means
)

}
