% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reducers.R
\name{return_if_single}
\alias{return_if_single}
\title{Reduce targets if more than one variable, return otherwise}
\usage{
return_if_single(.x, .f, ...)
}
\arguments{
\item{.x}{a \code{data.frame} containing the targets to reduce}

\item{.f}{a reduction function to apply}

\item{...}{arguments passed to \code{.f}}
}
\value{
a numeric vector, the reduced or original variable
}
\description{
Reduce targets if more than one variable, return otherwise
}
\keyword{internal}
