% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reducers.R
\name{rewind_target}
\alias{rewind_target}
\title{Set target to last value}
\usage{
rewind_target(.partition_step)
}
\arguments{
\item{.partition_step}{a \code{partition_step} object}
}
\value{
a \code{partition_step} object
}
\description{
Set target to last value
}
\keyword{internal}
