% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reducers.R
\name{scaled_mean}
\alias{scaled_mean}
\title{Average and scale rows in a \code{data.frame}}
\usage{
scaled_mean(.x, method = c("r", "c"))
}
\arguments{
\item{.x}{a \code{data.frame}}

\item{method}{The method source: both the pure R and C++ versions are efficient}
}
\value{
a numeric vector
}
\description{
\code{scaled_mean()} calculates scaled row means for a dataframe.
}
\examples{
library(dplyr)
iris \%>\%
  select_if(is.numeric) \%>\%
  scaled_mean()

}
