% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reducers.R
\name{search_k}
\alias{search_k}
\title{Search for the best \code{k}}
\usage{
search_k(.partition_step, search_method = c("binary", "linear"))
}
\arguments{
\item{.partition_step}{a \code{partition_step} object}

\item{search_method}{The search method. Binary search is generally more efficient
but linear search can be faster in very low dimensions.}
}
\value{
a \code{partition_step} object
}
\description{
Search for the best \code{k}
}
\keyword{internal}
