% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Count_data_simulate.R
\name{sim.nb}
\alias{sim.nb}
\title{Simulate NBZIN microbiome count data dependent on continuous covariate.}
\usage{
sim.nb(ntaxa, xvec, beta = 0, alt.ind = NULL)
}
\arguments{
\item{ntaxa}{Number of variables in the simulated microbiome dataset. Distribution
of these variables is negative binomial w/ zero inflation where the mean depends on
xvec through beta.}

\item{xvec}{A vector of values that influence the mean of the simulated data through beta
via a log-linear combination, that is, log(adjusted(mu)(i)) = mu.estimated + beta*xvec(i)*IQR(xvec),
for individual i, where IQR scales beta to the interquartile range of xvec.}

\item{beta}{Within each block, variables are correlated according to values
sampled from a uniform distribution with upper bound c.ub.}

\item{alt.ind}{Number of observations in the dataset.}
}
\description{
Data parameters were generated from genus summarized 16s rRNA count data
adjusted for read depth, excluding taxa with > 80 percent zeros.
Samples included > 20,000 reads
NBZIN parameters in dataset params estimated using the zeroinfl() function of the pscl R package.
}
\examples{
data(params)
ntaxa = 5
xvec = rnorm( 200 )
beta = .1
sim.nb( ntaxa, xvec, beta, alt.ind=1:ntaxa ) # all alternative hypotheses

for( beta in seq(.5,2, length.out=10) ){
	tmp = sim.nb( ntaxa, xvec, beta, alt.ind=1:ntaxa )
	tmpc = cor( tmp, method="spearman" )[1,2:ntaxa]
	print( paste("beta =", beta, "cor =", tmpc) )
	for( j in 1:ntaxa ) print( paste( cor(xvec,tmp[,j],method="spearman") ) )
}

}
