% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{simplify_names}
\alias{simplify_names}
\title{Simplify reduced variable names}
\usage{
simplify_names(.partition_step)
}
\arguments{
\item{.partition_step}{a \code{partition_step} object}
}
\value{
a \code{partition_step} object
}
\description{
Simplify reduced variable names
}
\keyword{internal}
