% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directors.R
\name{update_dist}
\alias{update_dist}
\title{Only fit the distances for a new variable}
\usage{
update_dist(.partition_step, spearman = FALSE)
}
\arguments{
\item{.partition_step}{a \code{partition_step} object}

\item{spearman}{Logical. Use Spearman's correlation?}
}
\value{
a \code{matrix}
}
\description{
Only fit the distances for a new variable
}
\keyword{internal}
