% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metric_min_r2}
\alias{metric_min_r2}
\title{Measure the information loss of reduction using minimum R-squared}
\usage{
metric_min_r2(.partition_step)
}
\arguments{
\item{.partition_step}{a \code{partition_step} object}
}
\value{
a \code{partition_step} object
}
\description{
Metrics are functions that tell how much information would be
lost for a given reduction in the data. reduce.  \code{\link[=as_metric]{as_metric()}} is a
helper function to create new metrics to be used in \code{partitioner}s.
\code{partitioner}s can be created with \code{\link[=as_partitioner]{as_partitioner()}}.

\code{metric_variance_explained()} assesses information loss by
calculating the minimum R-squared for the target variables.
}
