% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitioners.R
\name{part_kmeans}
\alias{part_kmeans}
\title{Partitioner: K-means, ICC, scaled means}
\usage{
part_kmeans()
}
\value{
a \code{partitioner}
}
\description{
\code{part_minr2()} uses the following functions:
\itemize{
\item \strong{director}: \code{direct_k_cluster()}, K-Means Clusters
\item \strong{metric}: \code{metric_min_icc()}, Minimum Intraclass Correlation
\item \strong{reducer}: \code{reduce_kmeans()}, Scaled Row Means

@md
}
}
