% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitioners.R
\name{part_minr2}
\alias{part_minr2}
\title{Partitioner: distance, minimum R-squared, scaled means}
\usage{
part_minr2()
}
\value{
a \code{partitioner}
}
\description{
\code{part_minr2()} uses the following functions:
\itemize{
\item \strong{director}: \code{direct_distance()}, Miniumum Distance
\item \strong{metric}: \code{metric_min_r2()}, Minimum R-Squared
\item \strong{reducer}: \code{reduce_scaled_mean()}, Scaled Row Means

@md
}
}
