% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitioners.R
\name{part_stdmi}
\alias{part_stdmi}
\title{Partitioner: distance, mutual information, scaled means}
\usage{
part_stdmi(spearman = FALSE)
}
\arguments{
\item{spearman}{logical. Use Spearman's correlation for distance matrix?}
}
\value{
a \code{partitioner}
}
\description{
\code{part_stdmi()} uses the following functions:
\itemize{
\item \strong{director}: \code{direct_distance()}, Miniumum Distance
\item \strong{metric}: \code{metric_std_mutualinfo()}, Standardized Mutual Information
\item \strong{reducer}: \code{reduce_scaled_mean()}, Scaled Row Means

@md
}
}
