\name{partition-package}
\alias{partition-package}
\docType{package}
\title{
 Agglomerative partitioning for dimension reduction
}
\description{
This package implements a formal statistic
}
\details{
\tabular{ll}{
Package: \tab cit\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2017-12-1\cr
License: \tab Artistic-2.0\cr
LazyLoad: \tab yes\cr
}
This package implements
}
\author{
Joshua Millstein

Maintainer: Joshua Millstein <joshua.millstein@usc.edu>
Joshua Millstein
}
\references{
Millstein J, 
}

\keyword{ dimension reduction }
\keyword{ clustering }
