% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\docType{package}
\name{partition}
\alias{partition}
\alias{partition-package}
\title{partition: Race to Solve the Partition Problem}
\description{
The partition package provides a collection of functions for a R student project.
The goal is to select the best algorithm for solving the partition problem with racing methods.
The partition problem consists in finding a bi-partition of a set of integers which minimizes the difference between their sums.
}

