% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{count_clusters}
\alias{count_clusters}
\alias{total_reduced}
\alias{summarize_mapping}
\alias{minimum_information}
\title{Helper functions to print \code{partition} summary}
\usage{
count_clusters(.partition)

total_reduced(.partition)

summarize_mapping(.partition, n_composite = 5, n_reduced = 10)

minimum_information(.partition, .round = TRUE, digits = 3)
}
\arguments{
\item{.partition}{a \code{partition} object}

\item{n_composite}{number of composite variables to print before summarizing}

\item{n_reduced}{number of reduced variables to print before summarizing}

\item{.round}{Should the minimum information be rounded?}

\item{digits}{If \code{.round} is \code{TRUE}, to what digit should it be rounded?}
}
\description{
Helper functions to print \code{partition} summary
}
\keyword{internal}
