% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reducers.R
\name{reduce_first_component}
\alias{reduce_first_component}
\title{Reduce selected variables to first principal component}
\usage{
reduce_first_component(.partition_step)
}
\arguments{
\item{.partition_step}{a \code{partition_step} object}
}
\value{
a \code{partition_step} object
}
\description{
Reducers are functions that tell the partition algorithm how
to reduce the data. \code{\link[=as_reducer]{as_reducer()}} is a helper function to create new
reducers to be used in \code{partitioner}s. \code{partitioner}s can be created with
\code{\link[=as_partitioner]{as_partitioner()}}.

\code{reduce_first_component()} returns the first component from the
principal components analysis of the target variables.
}
\seealso{
Other reducers: 
\code{\link{as_reducer}()},
\code{\link{reduce_kmeans}()},
\code{\link{reduce_scaled_mean}()}
}
\concept{reducers}
