% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Partition_Agglomerative.R
\name{sim_blk_diag_mvn}
\alias{sim_blk_diag_mvn}
\title{Simulate block diagonal multivariate normal data}
\usage{
sim_blk_diag_mvn(blk.vec, c.lb, c.ub, n, blk.names = "X.mvn", sep = ".")
}
\arguments{
\item{blk.vec}{Vector of block sizes that determines the total number of
variables as well as number of variables within each block.}

\item{c.lb}{Within each block, variables are correlated according to values
sampled from a uniform distribution with lower bound c.lb.}

\item{c.ub}{Within each block, variables are correlated according to values
sampled from a uniform distribution with upper bound c.ub.}

\item{n}{Number of observations in the dataset.}

\item{blk.names}{Character vector for column name prefix. Default is "X.mvn".}

\item{sep}{Character separating prefix and number in column names. Default is
".".}
}
\value{
A dataframe with sum(blk.vec) columns and n rows.
}
\description{
The MASS R package with function mvrnorm is used to simulate block diagonal
multivariate normal data, with correlations uniformly distributed within
blocks.
}
\details{
The MASS R package with function mvrnorm is used to simulate block diagonal
multivariate normal data, with correlations uniformly distributed within
blocks. Correlations within each block are randomly sampled from a U(c.lb,
c.ub) distribution prior to data generation.
}
\examples{


block_sizes <- 2:20
lower_bound <- .2
upper_bound <- .4
n <- 200

dat <- sim_blk_diag_mvn(block_sizes, lower_bound, upper_bound, n)

}
\references{
Millstein J, et al.
}
\author{
Joshua Millstein
}
